+++ CHANGELOG.md
@@ -7,5 +7,6 @@ All notable changes to this project will be documented in this file.
 - Added: support for utils bundle v3
 - Changed: require at least php 8.1
 - Changed: require at least contao 4.13
-- Deleted: deprecated classes from version 1
-- Deleted: huh.encore config key
+- Removed: deprecated classes from version 1
+- Removed: huh.encore config key
+- Removed: DcaGenerator class (use EncoreEntriesSelectField instead)